chore(META): remove code generation for most and rxjs packages
