+++ .scripts/RELEASABLE_PACKAGES
@@ -4,8 +4,8 @@ rxjs-run
 jsonp
 isolate
 time
-time/rxjs
-time/most
+rxjs-time
+most-time
 dom
 html
 http
