+++ Makefile
@@ -1,7 +1,6 @@
 .PHONY: lint docs dom history html http isolate jsonp most-run run rxjs-run
 
-ROOT_DIR:=$(shell dirname $(realpath $(lastword $(MAKEFILE_LIST))))
-BINDIR=$(ROOT_DIR)/node_modules/.bin
+BINDIR=node_modules/.bin
 TSLINT=$(BINDIR)/tslint
 TSC=$(BINDIR)/tsc
 MOCHA=$(BINDIR)/mocha
@@ -76,12 +75,10 @@ lib :
 			make lib $$d ; \
 		done < .scripts/RELEASABLE_PACKAGES; \
 	else \
-		dir=$(ROOT_DIR)/$(ARG) ;\
-		rm -rf $$dir/lib/ ;\
-		mkdir -p $$dir/lib ;\
-		grep 'prelib' $$dir/package.json >/dev/null && cd $$dir && npm run prelib ;\
-		$(TSC) --project $$dir --module commonjs --outDir $$dir/lib ;\
-		$(TSC) --project $$dir --module es6 --outDir $$dir/lib/es6 ;\
+		rm -rf $(ARG)/lib/ ;\
+		mkdir -p $(ARG)/lib ;\
+		$(TSC) --project $(ARG) --module commonjs --outDir $(ARG)/lib ;\
+		$(TSC) --project $(ARG) --module es6 --outDir $(ARG)/lib/es6 ;\
 		echo "✓ Compiled TypeScript to lib\n" ;\
 	fi
