+++ most-time/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../../tsconfig.common.json",
+  "extends": "../tsconfig.common.json",
   "compilerOptions": {
     "lib": [
       "dom",
