+++ rxjs-time/README.md
@@ -1,0 +1,3 @@
+## Only auxiliary package
+
+Please refer to [the main README](https://github.com/cyclejs/cyclejs/tree/master/time/README.md)
