+++ rxjs-time/tsconfig.json
@@ -1,0 +1,16 @@
+{
+  "extends": "../tsconfig.common.json",
+  "compilerOptions": {
+    "lib": [
+      "dom",
+      "es5",
+      "scripthost",
+      "es2015.collection",
+      "es2015.promise",
+      "es2015.iterable"
+    ]
+  },
+  "files": [
+    "src/index.ts"
+  ]
+}
