--- time/packageTemplate/README.md
@@ -1,3 +1,0 @@
-## Only auxiliary package
-
-Please refer to [the main README](https://github.com/cyclejs/cyclejs/tree/master/time/common/README.md)
