--- time/src/index.common.ts
@@ -1,5 +1,0 @@
-export {timeDriver as timeDriverUntyped} from './time-driver';
-export {mockTimeSource as mockTimeSourceUntyped} from './mock-time-source';
-export {Frame} from './animation-frames';
-export {Scheduler} from './scheduler';
-export * from './types';
